requirements.txt
torch>=1.7.1
torchvision>=0.8.2
numpy>=1.19.0
Pillow>=8.0.0
opencv-python>=4.5.1
scikit-learn>=0.24.0

git+https://github.com/openai/CLIP.git


matplotlib>=3.3.0
argparse>=1.4.0
tqdm>=4.50.0


jupyter>=1.0.0
ipykernel>=6.0.0